requirements.txt
Flask==2.1.2
python-snappy==0.6.1
requests==2.28.0
snappy-manifolds==1.1.2
protobuf==4.21.1